Update for ResponseWriter, use new finish function from  ResponseBodyWriter
